+++ Release_Notes.md
@@ -19,6 +19,7 @@
 - Renamed UploadResponse.cs to DTO_UploadResponse.cs
 - Moved MLogging.cs to Table/MLogging.cs
 - Added the GWCommon class to the Growthware.Framework.csproj to help with common functions
+- Moved the "Chunk Size" from the typescript to the configuration file and it is now retrieved from the API
 
 ### Known Bugs
 - There is an issue where the refresh token is being revoked when it shouldn't be causing a logout to occure